Remove reat import
