+++ hundredproof-react.code-workspace
@@ -3,5 +3,8 @@
 		{
 			"path": "."
 		}
-	]
+	],
+	"settings": {
+		"editor.formatOnSave": true
+	}
 }
