+++ src/components/AmountSelector/AmountSelector.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+
 
 import { DilutionContext } from '../../contexts/DilutionContext';
 import Input from '../Input/Input';
