+++ src/components/Button/Button.tsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import React from 'react';
+
 
 import styles from './Button.module.scss';
