+++ src/components/Footer/Footer.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+
 
 import styles from './Footer.module.scss';
