+++ src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
+
+import Link from 'next/link'
 
 import styles from './Header.module.scss';
 import Logo from './Logo';
