+++ src/components/Header/Logo.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+
 
 const Logo = () => {
   return (
