+++ src/components/Header/Subheader.tsx
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import React from 'react';
+
 
 import styles from './Header.module.scss';
