+++ src/components/IconSet/IconSet.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+
 
 const IconSet = () => {
   return (
