+++ src/components/Ingredients/IngredientItem.tsx
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import React from 'react';
+
 
 import Icon from '../Icon/Icon';
 import Input from '../Input/Input';
