+++ src/components/Input/Input.tsx
@@ -2,7 +2,7 @@
 
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import React from 'react';
+
 import AutosizeInput from 'react-input-autosize';
 
 import styles from './Input.module.scss';
