+++ src/components/MeasureHeader/MeasureHeader.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+
 
 import { DilutionContext } from '../../contexts/DilutionContext';
 import SegmentedButton from '../SegmentedButton/SegmentedButton';
