+++ src/contexts/ScalingContext.ts
@@ -1,3 +1,3 @@
-import React from 'react';
+
 
 export const ScalingContext = React.createContext();
