+++ src/utils/convertABV.ts
@@ -1,3 +1,3 @@
-export default function convertABV(measure, value) {
+export default function convertABV(measure: string, value: number) {
   return measure === 'proof' ? value * 2 : value / 2;
 }
