+++ src/utils/convertCups.ts
@@ -1,6 +1,6 @@
 import conversions from '../consts/conversions';
 import round from './round';
 
-export default function convertCups(amount) {
+export default function convertCups(amount: number) {
   return round(amount / conversions.ozToCups, 2);
 }
