+++ src/utils/convertTeaspoons.ts
@@ -1,6 +1,6 @@
 import conversions from '../consts/conversions';
 import round from './round';
 
-export default function convertTeaspoons(amount) {
+export default function convertTeaspoons(amount: number) {
   return round(amount / conversions.ozToTeaspoons, 2);
 }
